test: Prize 테스트 추가
